pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qasystem-on-medical-kg-huzi"
version = "0.1.0"
description = "医疗知识图谱问答系统"
requires-python = ">=3.13.2"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "chromadb>=1.1.1",
    "dotenv>=0.9.9",
    "neo4j>=6.0.2",
    "pandas>=2.3.3",
    "pyahocorasick>=2.2.0",
    "tqdm>=4.67.1",
    "sentence-transformers>=2.0.0",
    "openai>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
